Replaced stim null circuit on both headstages
- ADG140x was replaced with dg509b, which is a 2 circuit 1-to-4 mux.
This allows the electrode path to be completely disconnected from
stimulation circuitry during idle periods. During idle periods,
stimuluation output is shunted to ground through a dummy load so that
offsets to not develop large voltages during idle periods.
- Additionally, the inverting node of the estim current source amplifier
is sent to the intan chip instead of the DAC voltage. This gives a true
reflection of delivered current rather than just the control signal.